Refactor Configuration class
This commit introduces the following changes to the codebase:

- Added AppConfiguration class to enable or disable the use of fake repositories for products and customers separately.
- Removed unnecessary ProductConfiguration class.
- Updated application.properties to set 'app.useFakeProductRepository' and 'app.useFakeCustomerRepository' to 'false' by default, allowing more flexibility in choosing repository implementations.

Purpose:
The purpose of this commit is to enhance the flexibility of repository configuration. Now, we can independently control the use of fake repositories for products and customers, providing a more fine-grained approach to repository selection based on project's needs.
