+++ src/main/java/com/example/springboot/customer/CustomerService.java
@@ -11,7 +11,7 @@ public class CustomerService {
   private final CustomerRepository customerRepository;
 
   @Autowired
-  public CustomerService(@Qualifier("customerRepositoryFake") CustomerRepository customerRepository) {
+  public CustomerService(CustomerRepository customerRepository) {
     this.customerRepository = customerRepository;
   }
