+++ src/main/resources/application.properties
@@ -1,4 +1,5 @@
-app.useFakeProductRepository=true
+app.useFakeProductRepository=false
+app.useFakeCustomerRepository=false
 
 # Logging Configuration
